data_pipeline/requirements.txt
requests>=2.32.0
PyYAML>=6.0.1
pandas>=2.2.2
pandera>=0.20.3
snowflake-connector-python[pandas]>=3.10.0
python-dotenv>=1.0.1
tqdm>=4.66.4
